Initialize scroll metrics on page load
